client side can retrieve userid, and users cannot go to login/signup pages after signed in
