+++ types/next-auth.d.ts
@@ -1,0 +1,9 @@
+import NextAuth, {DefaultSession} from "next-auth"
+
+declare module "next-auth" {
+    interface Session {
+        user: {
+            userid: string
+        } & DefaultSession["user"]
+    }
+}
